fixed the clear
