--- ContactSheet.h
@@ -1,23 +1,0 @@
-#ifndef CONTACTSHEET_H
-#define CONTACTSHEET_H
-
-#include <QMainWindow>
-
-QT_BEGIN_NAMESPACE
-namespace Ui { class MainWindow; }
-QT_END_NAMESPACE
-
-class MainWindow : public QMainWindow
-{
-    Q_OBJECT
-
-public:
-    MainWindow(QWidget *parent = nullptr);
-    ~MainWindow();
-
-private slots:
-
-private:
-    Ui::MainWindow *ui;
-};
-#endif // CONTACTSHEET_H
