--- ContactSheet.ui
@@ -1,42 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ui version="4.0">
- <class>MainWindow</class>
- <widget class="QMainWindow" name="MainWindow">
-  <property name="geometry">
-   <rect>
-    <x>0</x>
-    <y>0</y>
-    <width>800</width>
-    <height>600</height>
-   </rect>
-  </property>
-  <property name="windowTitle">
-   <string>MainWindow</string>
-  </property>
-  <widget class="QWidget" name="centralwidget"/>
-  <widget class="QMenuBar" name="menubar">
-   <property name="geometry">
-    <rect>
-     <x>0</x>
-     <y>0</y>
-     <width>800</width>
-     <height>26</height>
-    </rect>
-   </property>
-   <widget class="QMenu" name="Information">
-    <property name="title">
-     <string>File</string>
-    </property>
-   </widget>
-   <addaction name="Information"/>
-  </widget>
-  <widget class="QStatusBar" name="statusbar"/>
-  <action name="actionInformation">
-   <property name="text">
-    <string>New window</string>
-   </property>
-  </action>
- </widget>
- <resources/>
- <connections/>
-</ui>
