+++ main.cpp
@@ -1,11 +1,11 @@
-#include "ContactSheet.h"
+#include "contact.h"
 
 #include <QApplication>
 
 int main(int argc, char *argv[])
 {
     QApplication a(argc, argv);
-    MainWindow w;
+    contact w;
     w.show();
     return a.exec();
 }
