--- mainwindow.h
@@ -1,26 +1,0 @@
-#ifndef MAINWINDOW_H
-#define MAINWINDOW_H
-
-#include <QMainWindow>
-
-QT_BEGIN_NAMESPACE
-namespace Ui { class MainWindow; }
-QT_END_NAMESPACE
-
-class MainWindow : public QMainWindow
-{
-    Q_OBJECT
-
-public:
-    MainWindow(QWidget *parent = nullptr);
-    ~MainWindow();
-
-public slots:
-    void displayAdmin();
-    void displayHome();
-    void displaySort();
-
-private:
-    Ui::MainWindow *ui;
-};
-#endif // MAINWINDOW_H
